refactor: move hash module to util folder
